hw5-auth done

+++ src/services/contacts.js
@@ -1,4 +1,4 @@
-import { Contact } from '../db/models/contactModel.js';
+import { Contact } from '../db/models/contact.js';
 import { calculatePaginationData } from '../utils/calculatePaginationData.js';
 import { SORT_ORDER } from '../constants/index.js';
 
@@ -9,11 +9,12 @@ export const getAllContacts = async ({
   sortOrder = SORT_ORDER.ASC,
   contactType,
   isFavourite,
+  userId,
 }) => {
   const skip = (page - 1) * perPage;
   const limit = perPage;
 
-  const filter = {};
+  const filter = { userId };
   if (contactType) filter.contactType = contactType;
   if (isFavourite !== undefined) filter.isFavourite = isFavourite === 'true';
 
@@ -31,13 +32,16 @@ export const getAllContacts = async ({
   };
 };
 
-export const getContactById = async (id) => await Contact.findById(id).lean();
+export const getContactById = async (id, userId) =>
+  await Contact.findById({ _id: id, userId }).lean();
 
 export const createContact = async (contactData) =>
   await Contact.create(contactData);
 
-export const updateContactById = async (contactId, updateData) =>
-  await Contact.findByIdAndUpdate(contactId, updateData, { new: true });
+export const updateContactById = async (contactId, updateData, userId) =>
+  await Contact.findByIdAndUpdate({ _id: contactId, userId }, updateData, {
+    new: true,
+  });
 
-export const deleteContactById = async (contactId) =>
-  await Contact.findByIdAndDelete(contactId);
+export const deleteContactById = async (contactId, userId) =>
+  await Contact.findByIdAndDelete({ _id: contactId, userId });
